drf/requirements/base.txt
Django==1.11.11
djangorestframework==3.6.2
django-rest-framework-social-oauth2==1.1.0
rest-condition==1.0.3
django-storages==1.6.6
boto==2.48.0
Pillow==5.1.0